Move database plugin to ./plugins
